+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/dsl/TimeoutImageEditReplacePatchable.java
@@ -16,7 +16,14 @@
 
 package io.fabric8.kubernetes.client.dsl;
 
+import java.util.concurrent.TimeUnit;
+
 public interface TimeoutImageEditReplacePatchable<T> extends
-    Timeoutable<ImageEditReplacePatchable<T>>,
-    ImageEditReplacePatchable<T> {
+    Timeoutable, ImageEditReplacePatchable<T> {
+
+  @Override
+  ImageEditReplacePatchable<T> withTimeout(long timeout, TimeUnit unit);
+
+  @Override
+  ImageEditReplacePatchable<T> withTimeoutInMillis(long timeoutInMillis);
 }