+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/extension/ResourceAdapter.java
@@ -26,6 +26,7 @@ import io.fabric8.kubernetes.client.ResourceNotFoundException;
 import io.fabric8.kubernetes.client.Watch;
 import io.fabric8.kubernetes.client.Watcher;
 import io.fabric8.kubernetes.client.dsl.Deletable;
+import io.fabric8.kubernetes.client.dsl.DeletableWithOptions;
 import io.fabric8.kubernetes.client.dsl.Gettable;
 import io.fabric8.kubernetes.client.dsl.Informable;
 import io.fabric8.kubernetes.client.dsl.NonDeletingOperation;
@@ -322,6 +323,16 @@ public class ResourceAdapter<T> implements Resource<T> {
   @Override
   public T serverSideApply() {
     return resource.serverSideApply();
+  }
+
+  @Override
+  public DeletableWithOptions withTimeout(long timeout, TimeUnit unit) {
+    return resource.withTimeout(timeout, unit);
+  }
+
+  @Override
+  public DeletableWithOptions withTimeoutInMillis(long timeoutInMillis) {
+    return withTimeout(timeoutInMillis, TimeUnit.MILLISECONDS);
   }
 
 }