+++ kubernetes-itests/src/test/java/io/fabric8/kubernetes/DeleteIT.java
@@ -17,14 +17,21 @@ package io.fabric8.kubernetes;
 
 import io.fabric8.junit.jupiter.api.LoadKubernetesManifests;
 import io.fabric8.kubernetes.api.model.DeletionPropagation;
+import io.fabric8.kubernetes.api.model.Secret;
+import io.fabric8.kubernetes.api.model.StatusDetails;
 import io.fabric8.kubernetes.api.model.apps.Deployment;
 import io.fabric8.kubernetes.api.model.apps.ReplicaSet;
 import io.fabric8.kubernetes.api.model.apps.ReplicaSetList;
 import io.fabric8.kubernetes.client.KubernetesClient;
+import io.fabric8.kubernetes.client.dsl.Resource;
+import io.fabric8.kubernetes.client.utils.CommonThreadPool;
 import org.junit.jupiter.api.Test;
 
+import java.util.List;
 import java.util.Objects;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -47,14 +54,44 @@ class DeleteIT {
   }
 
   @Test
-  void testDeleteExistentResource() {
+  void testDeleteNonExistentResourceBlocking() {
+    // Given
+    String podName = "i-dont-exist";
+    // When
+    long start = System.currentTimeMillis();
+    client.pods().withName(podName).withTimeout(10, TimeUnit.MINUTES).delete();
+    // Then it shouldn't block
+    assertTrue(System.currentTimeMillis() - start < TimeUnit.MILLISECONDS.convert(1, TimeUnit.MINUTES));
+  }
+
+  @Test
+  void testBlockingDeleteExistentResource() throws Exception {
     // Given
     String name = "deleteit-existent";
     // When
-    boolean isDeleted = client.secrets().withName(name).delete().size() == 1;
-    // Then
-    assertTrue(isDeleted);
-    client.secrets().withName(name).waitUntilCondition(Objects::isNull, 60, TimeUnit.SECONDS);
+    Resource<Secret> op = client.secrets().withName(name);
+    CompletableFuture<List<StatusDetails>> future;
+    try {
+      op.edit(s -> {
+        s.addFinalizer("fabric8.com/blocking");
+        return s;
+      });
+      future = CompletableFuture.supplyAsync(() -> op.withTimeout(60, TimeUnit.SECONDS).delete(), CommonThreadPool.get());
+
+      try {
+        future.get(10, TimeUnit.SECONDS);
+      } catch (TimeoutException e) {
+        // expected - we're waiting for the finalizer
+      }
+      assertNotNull(op.get().getMetadata().getDeletionTimestamp());
+    } finally {
+      op.edit(s -> {
+        s.getMetadata().setFinalizers(null);
+        return s;
+      });
+    }
+
+    assertEquals(1, future.get(60, TimeUnit.SECONDS).size());
   }
 
   @Test